src/api/requirements.txt
sqlmodel==0.0.18
fastapi==0.111.0
python-dotenv==1.0.1
alembic==1.13.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
requests==2.31.0
pandas==2.2.2